# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RedForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectWithStats"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /projects/{id}/scan:
    post:
      operationId: triggerScan
      tags: [projects]
      summary: Trigger a scan for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Scan started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
...
  /findings:
    get:
      operationId: listFindings
      tags: [findings]
      summary: List findings
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [CRITICAL, HIGH, MEDIUM, LOW, INFO]
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, IN_PROGRESS, FIXED, WONT_FIX, FALSE_POSITIVE]
        - name: projectId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of findings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FindingWithProject"
  /findings/{id}:
    get:
      operationId: getFinding
      tags: [findings]
      summary: Get a finding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Finding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateFinding
      tags: [findings]
      summary: Update finding status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFindingRequest"
      responses:
        "200":
          description: Finding updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
  /findings/{id}/generate-fix:
    post:
      operationId: generateFix
      tags: [findings]
      summary: Generate AI fix for a finding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Fix generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        targetUrl:
          type: string
        targetType:
          type: string
        status:
          type: string
        workspaceId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - targetUrl
        - targetType
        - status
        - workspaceId
        - createdAt
        - updatedAt
    ProjectWithStats:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        targetUrl:
          type: string
        targetType:
          type: string
        status:
          type: string
        workspaceId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastScanAt:
          type: string
          format: date-time
          nullable: true
        scanCount:
          type: integer
        findingCount:
          type: integer
        criticalCount:
          type: integer
      required:
        - id
        - name
        - targetUrl
        - targetType
        - status
        - workspaceId
        - createdAt
        - updatedAt
        - scanCount
        - findingCount
        - criticalCount
    ProjectDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        targetUrl:
          type: string
        targetType:
          type: string
        status:
          type: string
        workspaceId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        scans:
          type: array
          items:
            $ref: "#/components/schemas/ScanSummary"
        findings:
          type: array
          items:
            $ref: "#/components/schemas/FindingWithProject"
      required:
        - id
        - name
        - targetUrl
        - targetType
        - status
        - workspaceId
        - createdAt
        - updatedAt
        - scans
        - findings
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        targetUrl:
          type: string
        targetType:
          type: string
          enum: [WEB_APP, API, MOBILE_API, GRAPHQL]
      required:
        - name
        - targetUrl
        - targetType
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        targetUrl:
          type: string
        status:
          type: string
    Scan:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, FAILED]
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        findingsCount:
          type: integer
        criticalCount:
          type: integer
        highCount:
          type: integer
        mediumCount:
          type: integer
        lowCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - status
        - findingsCount
        - criticalCount
        - highCount
        - mediumCount
        - lowCount
        - createdAt
    ScanSummary:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        findingsCount:
          type: integer
        criticalCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - findingsCount
        - criticalCount
        - createdAt
...
    Finding:
      type: object
      properties:
        id:
          type: string
        scanId:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
        endpoint:
          type: string
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW, INFO]
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, FIXED, WONT_FIX, FALSE_POSITIVE]
        cvss:
          type: number
          nullable: true
        cwe:
          type: string
          nullable: true
        owasp:
          type: string
          nullable: true
        pocCode:
          type: string
          nullable: true
        fixPatch:
          type: string
          nullable: true
        fixExplanation:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - scanId
        - projectId
        - title
        - description
        - endpoint
        - severity
        - status
        - createdAt
        - updatedAt
    FindingWithProject:
      type: object
      properties:
        id:
          type: string
        scanId:
          type: string
        projectId:
          type: string
        projectName:
          type: string
        title:
          type: string
        description:
          type: string
        endpoint:
          type: string
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW, INFO]
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, FIXED, WONT_FIX, FALSE_POSITIVE]
        cvss:
          type: number
          nullable: true
        cwe:
          type: string
          nullable: true
        owasp:
          type: string
          nullable: true
        pocCode:
          type: string
          nullable: true
        fixPatch:
          type: string
          nullable: true
        fixExplanation:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - scanId
        - projectId
        - projectName
        - title
        - description
        - endpoint
        - severity
        - status
        - createdAt
        - updatedAt
    UpdateFindingRequest:
      type: object
      properties:
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, FIXED, WONT_FIX, FALSE_POSITIVE]
...